Move selectors to loading, entity loading/error domains
