+++ src/components/AppCatalog/AppList/AppList.js
@@ -8,10 +8,8 @@ import { Link } from 'react-router-dom';
 import { AppCatalogRoutes } from 'shared/constants/routes';
 import { listCatalogs } from 'stores/appcatalog/actions';
 import { CATALOG_LOAD_INDEX_REQUEST } from 'stores/appcatalog/constants';
-import {
-  selectLoadingFlagByAction,
-  selectLoadingFlagByIdAndAction,
-} from 'stores/cluster/selectors';
+import { selectLoadingFlagByIdAndAction } from 'stores/entityloading/selectors';
+import { selectLoadingFlagByAction } from 'stores/loading/selectors';
 import LoadingOverlay from 'UI/LoadingOverlay';
 
 import AppListInner from './AppListInner';
