+++ src/components/Cluster/ClusterDetail/ClusterApps.js
@@ -8,11 +8,9 @@ import { connect } from 'react-redux';
 import { Constants } from 'shared/constants';
 import { AppCatalogRoutes } from 'shared/constants/routes';
 import { loadClusterApps } from 'stores/appcatalog/actions';
-import {
-  selectClusterById,
-  selectErrorByIdAndAction,
-} from 'stores/cluster/selectors';
+import { selectClusterById } from 'stores/cluster/selectors';
 import { isClusterCreating, isClusterUpdating } from 'stores/cluster/utils';
+import { selectErrorByIdAndAction } from 'stores/entityerror/selectors';
 import { selectCluster } from 'stores/global/actions';
 import Button from 'UI/Button';
 import ClusterDetailPreinstalledApp from 'UI/ClusterDetailPreinstalledApp';
