+++ src/stores/entityloading/selectors.ts
@@ -1,0 +1,12 @@
+import { IState } from 'reducers/types';
+import { typeWithoutSuffix } from 'selectors/selectorUtils';
+
+export function selectLoadingFlagByIdAndAction(
+  state: IState,
+  id: string,
+  actionType: string
+) {
+  return (
+    state.loadingFlagsByEntity[id]?.[typeWithoutSuffix(actionType)] ?? true
+  );
+}
